relocate apollo provider and user jwt sourcing
