+++ packages/app-extension/src/components/Unlocked/index.tsx
@@ -1,4 +1,7 @@
-import { useBootstrapFast } from "@coral-xyz/recoil";
+import { ApolloProvider, SuspenseCache } from "@apollo/client";
+import { createApolloClient } from "@coral-xyz/common";
+import { Loading } from "@coral-xyz/react-common";
+import { useAuthenticatedUser, useBootstrapFast } from "@coral-xyz/recoil";
 
 import { Spotlight } from "../../spotlight/Spotlight";
 import { Router } from "../common/Layout/Router";
@@ -9,30 +12,41 @@ import { ApproveTransactionRequest } from "./ApproveTransactionRequest";
 import { PrimaryPubkeySelector } from "./PrimaryPubkeySelector";
 import { WithVersion } from "./WithVersion";
 
+const suspenseCache = new SuspenseCache();
+
 //
 // The main nav persistent stack.
 //
 export function Unlocked() {
   useBootstrapFast();
 
+  const user = useAuthenticatedUser();
+  if (!user) {
+    return <Loading />;
+  }
+
+  const apolloClient = createApolloClient(user.jwt);
+
   return (
-    <WithVersion>
-      <WalletDrawerProvider>
-        <Spotlight />
-        <WithTabs>
-          <div
-            style={{
-              display: "flex",
-              flexDirection: "column",
-              height: "100%",
-            }}
-          >
-            <Router />
-            <ApproveTransactionRequest />
-            <PrimaryPubkeySelector />
-          </div>
-        </WithTabs>
-      </WalletDrawerProvider>
-    </WithVersion>
+    <ApolloProvider client={apolloClient} suspenseCache={suspenseCache}>
+      <WithVersion>
+        <WalletDrawerProvider>
+          <Spotlight />
+          <WithTabs>
+            <div
+              style={{
+                display: "flex",
+                flexDirection: "column",
+                height: "100%",
+              }}
+            >
+              <Router />
+              <ApproveTransactionRequest />
+              <PrimaryPubkeySelector />
+            </div>
+          </WithTabs>
+        </WalletDrawerProvider>
+      </WithVersion>
+    </ApolloProvider>
   );
 }
